requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pywebview>=6.0.0
